+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- A new `metadata` property to the element clicked iframe message which contains element rect.
+
 ## [1.1.1] - 2020-10-01
 
 ### Added